requirements.txt
django==1.11.2
djangorestframework==3.5.4
django-api-commons==1.0.2
social-auth-core==1.4.0
social-auth-app-django==1.2.0
typing